+++ tests/acceptance/bashunit_assert_subcommand_test.sh
@@ -1,10 +1,6 @@
 #!/usr/bin/env bash
 set -euo pipefail
 
-function set_up_before_script() {
-  TEST_ENV_FILE="tests/acceptance/fixtures/.env.default"
-}
-
 function set_up() {
   export BASHUNIT_SIMPLE_OUTPUT=false
 }